Split CLI build graph targets
